chore: added new meta endpoint
